Seperated Policy and Agent
