+++ src/ReinforcementLearning/SimpleAgent/BaseAgent.h
@@ -2,23 +2,11 @@
 #define BASE_AGENT_H
 
 #include "../../alignment/State.h"
-
-/** \brief This Policy class maps states to an action deterministically.
-*/
-class Policy {
-    public:
-    Policy(){};
-    ~Policy(){};
-    /** Member policy selects edges in order as long as they are selectable.
-     * \param s Expects a state as input parameter.
-     * \return Returns the index of edge that was selected according to policy.
-     */
-    virtual unsigned int act(state* s) const = 0; 
-};
+#include "Policy.h"
 
 /** \brief This BaseAgent class executes Policy and saves the results.
 */
-class BaseAgent : public Policy {
+class BaseAgent {
     public:
     BaseAgent(){};
     ~BaseAgent(){};
