+++ src/ReinforcementLearning/SimpleAgent/Makefile
@@ -11,7 +11,7 @@ all: RandomAgent_main
 RandomAgent_main: $(OBJ)
 	g++ $(CXXFLAGS) $(OBJ) -o RandomAgent_main
 
-RandomAgent_main.o: RandomAgent_main.cc RandomAgent.h BaseAgent.h $(H1) $(H2) $(H3) $(H4)
+RandomAgent_main.o: RandomAgent_main.cc RandomAgent.h BaseAgent.h Policy.h RandomPolicy.h $(H1) $(H2) $(H3) $(H4)
 	g++ $(CXXFLAGS) -c RandomAgent_main.cc
 
 State.o: ../../alignment/State.cc $(H4) $(H2) $(H3)
